+++ test/svelte3/integration/bind/bind.multiple.svelte
@@ -1,0 +1,8 @@
+<input bind:value={number} />
+<PlusMinusControl bind:numberValue={number} />
+
+<script>
+    import PlusMinusControl from './PlusMinusControl.svelte';
+
+    let number;
+</script>