+++ test/svelte3/integration/bind/bind.named.svelte
@@ -1,5 +1,7 @@
 <ShopingCart bind:totalCost={cost} />
 
 <script>
+    let cost;
+    
     import ShopingCart from './ShopingCart.svelte';
 </script>